Add the ARP cache and wire it into the layer 3 send path
arp.c gains a lookup, a store and expiry. The cache lives in net_interface. Entries are keyed by address and carry a timestamp, used both to evict the oldest when full and to drop stale ones on every lookup.

arp_handle caches the sender pair from every ARP it sees, including requests aimed at other hosts, and replies only when the request targets our own address.

send_packet now resolves the next hop through the cache, sending a request and returning NET_PENDING on a miss.

recv_packet exists and takes payload bytes rather than pulling a frame itself.

net_init fills in the interface. Status codes across link.c and network.c now use their named constants throughout.

+++ hub.c
@@ -65,7 +65,7 @@ int main(int argc, char *argv[]) {
 		else {
 			int skip = find_port(connections, &sender, active_ports);
 			printf("Received and retransmitting\n");
-			for (size_t i = 0; i < active_ports; i++) {
+			for (int i = 0; i < active_ports; i++) {
 				if (i == skip) continue;
 				sendto(socket_fd, buf, received, 0, (struct sockaddr *) (connections+i), sizeof connections[i]);
 			}
